# specs/1-shopping-poc/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Shopping POC API
  version: "1.0.0"
servers:
  - url: http://localhost:8000
paths:
  /api/v1/products:
    get:
      summary: List products
      responses:
        "200":
          description: OK
    post:
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductIn"
      responses:
        "200":
          description: Created
  /api/v1/auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "200":
          description: Token
  /api/v1/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "200":
          description: Token
  /api/v1/orders:
    post:
      summary: Place order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreate"
      responses:
        "200":
          description: Order created
components:
  schemas:
    ProductIn:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        available:
          type: boolean
      required: [name, price]
    UserCreate:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    OrderItem:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required: [productId, quantity, price]
    OrderCreate:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required: [items]